added back callback.
Learned that just the 'if submit:' lines after the form do NOT solve the problem, it was all the callback, the good and bad.  So I'll work on that.

The callback seems to get me closer to being able to set the values I want, that aren't getting set without it.  Though they did in the non-form version, just there wasn't control.  Another option might be to see if you can get control in that domain.