Corrected error, noted bug
Added Bug to fix here comment for a potential bad path the user can cause by refreshing so the code never resets the values.  Still want to improve it because it leaves the user is a weird place with no pool or anagram but a success method.  They CAN reset and restart but its weird.  Should be remedied so that the aforementioned bad reset path ( see comment around line 30) resets and somehow restarts the process.  Possibly with a session state value and conditional flow.