Moved name filtering
Now og name is reset to the stripped version of the input if it contained digits or punctuation.  Before this commit, a name like Michae3*^@82L would be processed as Michael but if an anagram is found, the screen shows: <anagram> Is an anagram for Michae3*^@82L!  Which is weird.  So make the assumption that the user doesn't mean to include punctuation or digits or whitespace total, by including this aspect (the name display on success).